# Improved performance of iterator of `Utf8Array` and `BinaryArray` (3-4x)