```
iter_values 2^20        time:   [1.6717 ms 1.6744 ms 1.6774 ms]                              
                        change: [-75.148% -75.071% -74.997%] (p = 0.00 < 0.05)
iter 2^20               time:   [2.2270 ms 2.2301 ms 2.2336 ms]                       
                        change: [-66.754% -66.678% -66.604%] (p = 0.00 < 0.05)
```

This has implications on almost every kernel over these arrays, as well as users of the iterator.