@jorgecarleitao 2021-09-20
Good catch, @Dandandan , I also rebased incorrectly and removed the `simdutf8`, so great to have reviewed it again :)