@ritchie46 2021-09-20
Very nice perf. improvement! Could you explain what you did? I've been browsing a bit, but I can't find it yet. Is it this line? 

```
let slice = self.values.get_unchecked(start..end);
```